+++ apps/atl_benchmark/atl/blurpart.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void blurpart(u_int8_t* v,int m,int n,u_int8_t*output);
+void blurpart(float* v,int m,int n,float*output);
