+++ apps/atl_benchmark/atl/conv1.c
@@ -1,20 +1,20 @@
 #include <stdlib.h>
 #include "conv1.h"
 
-void conv1(int m,int n,u_int8_t* c,u_int8_t*output){
-	u_int8_t* tmp4 = (u_int8_t*) aligned_alloc(4,(n - ((-(m)) + 1)) * sizeof (u_int8_t));
+void conv1(int m,int n,float* c,float*output){
+	float* tmp4 = (float*) aligned_alloc(4,(n - ((-(m)) + 1)) * sizeof (float));
 	for (int H3 = 0; H3 < n; H3++) {
 		for (int H4 = 0; H4 < m; H4++) {
-			u_int8_t tmp2 = 0;
+			float tmp2 = 0;
 			if (0 <= H3 - (H4)) {
-				u_int8_t tmp3 = 0;
+				float tmp3 = 0;
 				for (int H9 = (-(m)) + 1; H9 < n; H9++) {
 					if (H9 == 2 + (-(m)) + 1) {
 						tmp4[(H9 - ((-(m)) + 1))] = 1;
 					}
 				}
 				tmp3 = tmp4[H3 - (H4)];
-				u_int8_t tmp5 = 0;
+				float tmp5 = 0;
 				tmp5 = c[H4];
 				tmp2 = tmp3 * tmp5;
 			}
