+++ apps/atl_benchmark/atl/conv4.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void conv4(int m,int n,u_int8_t* c,u_int8_t*output);
+void conv4(int m,int n,float* c,float*output);
