+++ apps/atl_benchmark/atl/fusetest.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void fusetest(int m,int n,u_int8_t* v,u_int8_t*output);
+void fusetest(int m,int n,float* v,float*output);
