+++ apps/atl_benchmark/atl/gather.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void gather(u_int8_t* w,u_int8_t* x,int C,int W,u_int8_t*output);
+void gather(float* w,float* x,int C,int W,float*output);
