+++ apps/atl_benchmark/atl/im2collifted.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void im2collifted(u_int8_t* x,u_int8_t* w,int RR,int W,int K,int B,int A,u_int8_t*output);
+void im2collifted(float* x,float* w,int RR,int W,int K,int B,int A,float*output);
