+++ apps/atl_benchmark/atl/scatter.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void scatter(u_int8_t* w,u_int8_t* x,int C,int W,u_int8_t*output);
+void scatter(float* w,float* x,int C,int W,float*output);
